requirements.txt
argparse
sounddevice
openai-whisper
soundfile
pyperclip
